% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppWrap-package.R
\docType{package}
\name{RcppWrap-package}
\alias{RcppWrap}
\alias{RcppWrap-package}
\title{RcppWrap: C++ & R Wrapping Methods}
\description{
Shows how to wrap C++ structs for portability back'n'forth between R and C++
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/r-pkg-examples/rcpp-custom-wrap-as}
}

}
\author{
\strong{Maintainer}: James Balamuta \email{balamut2@illinois.edu} [copyright holder]

}
